+++ Log-O-Matic/src/components.d.ts
@@ -11,5 +11,6 @@ declare module 'vue' {
     HelloWorld: typeof import('./components/HelloWorld.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    StudentCardComponent: typeof import('./components/StudentCardComponent.vue')['default']
   }
 }
